Windows: Better support PAGER settings with spaces in the path
On Windows, the pager was always run via a shell, which is not the case
on Unix, where the shell variant was only used as a fallback.

Thus, setting the pager, for example, like this:

  PAGER="C:/Program Files/msys/bin/less"

would fail; an extra set of quotes is needed:

  PAGER="\"C:/Program Files/msys/bin/less\""

With this patch the former setting works as well.

Signed-off-by: Johannes Sixt <johannes.sixt@telecom.at>